matrix ops refactor
  1. split the test case from matrixops, and add to MatrixTest.cc

  2. add parameter to quantized_mat_mul for representing whethere the
  matrix is transpose or not
